Move `getFluentBitLabels` to "correct" file
