+++ pkg/component/observability/logging/fluentoperator/fluentoperator.go
@@ -33,10 +33,8 @@ import (
 	resourcesv1alpha1 "github.com/gardener/gardener/pkg/apis/resources/v1alpha1"
 	"github.com/gardener/gardener/pkg/client/kubernetes"
 	"github.com/gardener/gardener/pkg/component"
-	valiconstants "github.com/gardener/gardener/pkg/component/observability/logging/vali/constants"
 	"github.com/gardener/gardener/pkg/resourcemanager/controller/garbagecollector/references"
 	"github.com/gardener/gardener/pkg/utils"
-	gardenerutils "github.com/gardener/gardener/pkg/utils/gardener"
 	"github.com/gardener/gardener/pkg/utils/managedresources"
 )
 
@@ -328,17 +326,5 @@ func getLabels() map[string]string {
 		v1beta1constants.LabelApp:   name,
 		v1beta1constants.LabelRole:  v1beta1constants.LabelLogging,
 		v1beta1constants.GardenRole: v1beta1constants.GardenRoleLogging,
-	}
-}
-
-func getFluentBitLabels() map[string]string {
-	return map[string]string{
-		v1beta1constants.LabelApp:                             v1beta1constants.DaemonSetNameFluentBit,
-		v1beta1constants.LabelRole:                            v1beta1constants.LabelLogging,
-		v1beta1constants.GardenRole:                           v1beta1constants.GardenRoleLogging,
-		v1beta1constants.LabelNetworkPolicyToDNS:              v1beta1constants.LabelNetworkPolicyAllowed,
-		v1beta1constants.LabelNetworkPolicyToRuntimeAPIServer: v1beta1constants.LabelNetworkPolicyAllowed,
-		gardenerutils.NetworkPolicyLabel(valiconstants.ServiceName, valiconstants.ValiPort): v1beta1constants.LabelNetworkPolicyAllowed,
-		"networking.resources.gardener.cloud/to-all-shoots-logging-tcp-3100":                v1beta1constants.LabelNetworkPolicyAllowed,
 	}
 }
